Outsource more into baseclass
