+++ slowquant/unitary_coupled_cluster/linear_response/lr_baseclass.py
@@ -1,9 +1,35 @@
+import copy
 from collections.abc import Sequence
 
 import numpy as np
 import scipy
 
+from slowquant.unitary_coupled_cluster.operator_hybrid import (
+    OperatorHybrid,
+    convert_pauli_to_hybrid_form,
+)
+from slowquant.unitary_coupled_cluster.operator_pauli import (
+    energy_hamiltonian_pauli,
+    epq_pauli,
+    hamiltonian_pauli_1i_1a,
+    hamiltonian_pauli_2i_2a,
+)
 from slowquant.unitary_coupled_cluster.ucc_wavefunction import WaveFunctionUCC
+from slowquant.unitary_coupled_cluster.util import ThetaPicker, construct_ucc_u
+
+
+class ResponseOperator:
+    def __init__(self, occ_idx: Sequence[int], unocc_idx: Sequence[int], operator: OperatorHybrid) -> None:
+        """Initialize response excitation operator.
+
+        Args:
+            occ_idx: Index of occupied orbitals.
+            unocc_idx: Index of unoccupied orbitals.
+            operator: Operator.
+        """
+        self.occ_idx = occ_idx
+        self.unocc_idx = unocc_idx
+        self.operator = operator
 
 
 class LinearResponseBaseClass:
@@ -13,8 +39,147 @@ class LinearResponseBaseClass:
     Delta: np.ndarray
     wf: WaveFunctionUCC
 
-    def __init__(self) -> None:
-        raise NotImplementedError
+    def __init__(
+        self,
+        wave_function: WaveFunctionUCC,
+        excitations: str,
+        is_spin_conserving: bool = False,
+    ) -> None:
+        """Initialize linear response by calculating the needed matrices.
+
+        Args:
+            wave_function: Wave function object.
+            excitations: Which excitation orders to include in response.
+            is_spin_conserving: Use spin-conseving operators.
+        """
+        self.wf = copy.deepcopy(wave_function)
+        self.theta_picker = ThetaPicker(
+            self.wf.active_occ_spin_idx,
+            self.wf.active_unocc_spin_idx,
+            is_spin_conserving=is_spin_conserving,
+        )
+
+        self.G_ops: list[ResponseOperator] = []
+        self.q_ops: list[ResponseOperator] = []
+        num_spin_orbs = self.wf.num_spin_orbs
+        num_elec = self.wf.num_elec
+        excitations = excitations.lower()
+
+        if "s" in excitations:
+            for _, a, i, op_ in self.theta_picker.get_t1_generator_sa(num_spin_orbs, num_elec):
+                op = convert_pauli_to_hybrid_form(
+                    op_,
+                    self.wf.num_inactive_spin_orbs,
+                    self.wf.num_active_spin_orbs,
+                    self.wf.num_virtual_spin_orbs,
+                )
+                self.G_ops.append(ResponseOperator((i,), (a,), op))
+        if "d" in excitations:
+            for _, a, i, b, j, op_ in self.theta_picker.get_t2_generator_sa(num_spin_orbs, num_elec):
+                op = convert_pauli_to_hybrid_form(
+                    op_,
+                    self.wf.num_inactive_spin_orbs,
+                    self.wf.num_active_spin_orbs,
+                    self.wf.num_virtual_spin_orbs,
+                )
+                self.G_ops.append(ResponseOperator((i, j), (a, b), op))
+        if "t" in excitations:
+            for _, a, i, b, j, c, k, op_ in self.theta_picker.get_t3_generator(num_spin_orbs, num_elec):
+                op = convert_pauli_to_hybrid_form(
+                    op_,
+                    self.wf.num_inactive_spin_orbs,
+                    self.wf.num_active_spin_orbs,
+                    self.wf.num_virtual_spin_orbs,
+                )
+                self.G_ops.append(ResponseOperator((i, j, k), (a, b, c), op))
+        if "q" in excitations:
+            for _, a, i, b, j, c, k, d, l, op_ in self.theta_picker.get_t4_generator(num_spin_orbs, num_elec):
+                op = convert_pauli_to_hybrid_form(
+                    op_,
+                    self.wf.num_inactive_spin_orbs,
+                    self.wf.num_active_spin_orbs,
+                    self.wf.num_virtual_spin_orbs,
+                )
+                self.G_ops.append(ResponseOperator((i, j, k, l), (a, b, c, d), op))
+        if "5" in excitations:
+            for _, a, i, b, j, c, k, d, l, e, m, op_ in self.theta_picker.get_t5_generator(
+                num_spin_orbs, num_elec
+            ):
+                op = convert_pauli_to_hybrid_form(
+                    op_,
+                    self.wf.num_inactive_spin_orbs,
+                    self.wf.num_active_spin_orbs,
+                    self.wf.num_virtual_spin_orbs,
+                )
+                self.G_ops.append(ResponseOperator((i, j, k, l, m), (a, b, c, d, e), op))
+        if "6" in excitations:
+            for _, a, i, b, j, c, k, d, l, e, m, f, n, op_ in self.theta_picker.get_t6_generator(
+                num_spin_orbs, num_elec
+            ):
+                op = convert_pauli_to_hybrid_form(
+                    op_,
+                    self.wf.num_inactive_spin_orbs,
+                    self.wf.num_active_spin_orbs,
+                    self.wf.num_virtual_spin_orbs,
+                )
+                self.G_ops.append(ResponseOperator((i, j, k, l, m, n), (a, b, c, d, e, f), op))
+        for i, a in self.wf.kappa_idx:
+            op_ = 2 ** (-1 / 2) * epq_pauli(a, i, self.wf.num_spin_orbs, self.wf.num_elec)
+            op = convert_pauli_to_hybrid_form(
+                op_,
+                self.wf.num_inactive_spin_orbs,
+                self.wf.num_active_spin_orbs,
+                self.wf.num_virtual_spin_orbs,
+            )
+            self.q_ops.append(ResponseOperator((i), (a), op))
+
+        num_parameters = len(self.G_ops) + len(self.q_ops)
+        self.A = np.zeros((num_parameters, num_parameters))
+        self.B = np.zeros((num_parameters, num_parameters))
+        self.Sigma = np.zeros((num_parameters, num_parameters))
+        self.Delta = np.zeros((num_parameters, num_parameters))
+        self.H_2i_2a = convert_pauli_to_hybrid_form(
+            hamiltonian_pauli_2i_2a(
+                self.wf.h_ao,
+                self.wf.g_ao,
+                self.wf.c_trans,
+                self.wf.num_inactive_spin_orbs,
+                self.wf.num_active_spin_orbs,
+                self.wf.num_virtual_spin_orbs,
+                num_elec,
+            ),
+            self.wf.num_inactive_spin_orbs,
+            self.wf.num_active_spin_orbs,
+            self.wf.num_virtual_spin_orbs,
+        )
+        self.H_1i_1a = convert_pauli_to_hybrid_form(
+            hamiltonian_pauli_1i_1a(
+                self.wf.h_ao,
+                self.wf.g_ao,
+                self.wf.c_trans,
+                self.wf.num_inactive_spin_orbs,
+                self.wf.num_active_spin_orbs,
+                self.wf.num_virtual_spin_orbs,
+                num_elec,
+            ),
+            self.wf.num_inactive_spin_orbs,
+            self.wf.num_active_spin_orbs,
+            self.wf.num_virtual_spin_orbs,
+        )
+        self.H_en = convert_pauli_to_hybrid_form(
+            energy_hamiltonian_pauli(
+                self.wf.h_ao,
+                self.wf.g_ao,
+                self.wf.c_trans,
+                self.wf.num_inactive_spin_orbs,
+                self.wf.num_active_spin_orbs,
+                self.wf.num_virtual_spin_orbs,
+                num_elec,
+            ),
+            self.wf.num_inactive_spin_orbs,
+            self.wf.num_active_spin_orbs,
+            self.wf.num_virtual_spin_orbs,
+        )
 
     def calc_excitation_energies(self) -> None:
         """Calculate excitation energies."""
