+++ slowquant/unitary_coupled_cluster/linear_response/projected.py
@@ -19,33 +19,11 @@ from slowquant.unitary_coupled_cluster.linear_response.lr_baseclass import (
     LinearResponseBaseClass,
 )
 from slowquant.unitary_coupled_cluster.operator_hybrid import (
-    OperatorHybrid,
-    OperatorHybridData,
     convert_pauli_to_hybrid_form,
     expectation_value_hybrid_flow,
 )
-from slowquant.unitary_coupled_cluster.operator_pauli import (
-    OperatorPauli,
-    energy_hamiltonian_pauli,
-    epq_pauli,
-    hamiltonian_pauli_1i_1a,
-)
+from slowquant.unitary_coupled_cluster.operator_pauli import OperatorPauli, epq_pauli
 from slowquant.unitary_coupled_cluster.ucc_wavefunction import WaveFunctionUCC
-from slowquant.unitary_coupled_cluster.util import ThetaPicker, construct_ucc_u
-
-
-class ResponseOperator:
-    def __init__(self, occ_idx: Sequence[int], unocc_idx: Sequence[int], operator: OperatorHybrid) -> None:
-        """Initialize response excitation operator.
-
-        Args:
-            occ_idx: Index of occupied orbitals.
-            unocc_idx: Index of unoccupied orbitals.
-            operator: Operator.
-        """
-        self.occ_idx = occ_idx
-        self.unocc_idx = unocc_idx
-        self.operator = operator
 
 
 class LinearResponseUCC(LinearResponseBaseClass):
@@ -62,136 +40,8 @@ class LinearResponseUCC(LinearResponseBaseClass):
             excitations: Which excitation orders to include in response.
             is_spin_conserving: Use spin-conseving operators.
         """
-        self.wf = copy.deepcopy(wave_function)
-        self.theta_picker = ThetaPicker(
-            self.wf.active_occ_spin_idx,
-            self.wf.active_unocc_spin_idx,
-            is_spin_conserving=is_spin_conserving,
-        )
+        super().__init__(wave_function, excitations, is_spin_conserving)
 
-        self.G_ops: list[ResponseOperator] = []
-        self.q_ops: list[ResponseOperator] = []
-        num_spin_orbs = self.wf.num_spin_orbs
-        num_elec = self.wf.num_elec
-        excitations = excitations.lower()
-        U_matrix = construct_ucc_u(
-            self.wf.num_active_spin_orbs,
-            self.wf.num_active_elec,
-            self.wf.theta1
-            + self.wf.theta2
-            + self.wf.theta3
-            + self.wf.theta4
-            + self.wf.theta5
-            + self.wf.theta6,
-            self.wf.singlet_excitation_operator_generator,
-            "sdtq56",  # self.wf._excitations,
-        )
-        inactive_str = "I" * self.wf.num_inactive_spin_orbs
-        virtual_str = "I" * self.wf.num_virtual_spin_orbs
-        self.U = OperatorHybrid(
-            {inactive_str + virtual_str: OperatorHybridData(inactive_str, U_matrix, virtual_str)}
-        )
-        if "s" in excitations:
-            for _, a, i, op_ in self.theta_picker.get_t1_generator_sa(num_spin_orbs, num_elec):
-                op = convert_pauli_to_hybrid_form(
-                    op_,
-                    self.wf.num_inactive_spin_orbs,
-                    self.wf.num_active_spin_orbs,
-                    self.wf.num_virtual_spin_orbs,
-                )
-                self.G_ops.append(ResponseOperator((i,), (a,), op))
-        if "d" in excitations:
-            for _, a, i, b, j, op_ in self.theta_picker.get_t2_generator_sa(num_spin_orbs, num_elec):
-                op = convert_pauli_to_hybrid_form(
-                    op_,
-                    self.wf.num_inactive_spin_orbs,
-                    self.wf.num_active_spin_orbs,
-                    self.wf.num_virtual_spin_orbs,
-                )
-                self.G_ops.append(ResponseOperator((i, j), (a, b), op))
-        if "t" in excitations:
-            for _, a, i, b, j, c, k, op_ in self.theta_picker.get_t3_generator(num_spin_orbs, num_elec):
-                op = convert_pauli_to_hybrid_form(
-                    op_,
-                    self.wf.num_inactive_spin_orbs,
-                    self.wf.num_active_spin_orbs,
-                    self.wf.num_virtual_spin_orbs,
-                )
-                self.G_ops.append(ResponseOperator((i, j, k), (a, b, c), op))
-        if "q" in excitations:
-            for _, a, i, b, j, c, k, d, l, op_ in self.theta_picker.get_t4_generator(num_spin_orbs, num_elec):
-                op = convert_pauli_to_hybrid_form(
-                    op_,
-                    self.wf.num_inactive_spin_orbs,
-                    self.wf.num_active_spin_orbs,
-                    self.wf.num_virtual_spin_orbs,
-                )
-                self.G_ops.append(ResponseOperator((i, j, k, l), (a, b, c, d), op))
-        if "5" in excitations:
-            for _, a, i, b, j, c, k, d, l, e, m, op_ in self.theta_picker.get_t5_generator(
-                num_spin_orbs, num_elec
-            ):
-                op = convert_pauli_to_hybrid_form(
-                    op_,
-                    self.wf.num_inactive_spin_orbs,
-                    self.wf.num_active_spin_orbs,
-                    self.wf.num_virtual_spin_orbs,
-                )
-                self.G_ops.append(ResponseOperator((i, j, k, l, m), (a, b, c, d, e), op))
-        if "6" in excitations:
-            for _, a, i, b, j, c, k, d, l, e, m, f, n, op_ in self.theta_picker.get_t6_generator(
-                num_spin_orbs, num_elec
-            ):
-                op = convert_pauli_to_hybrid_form(
-                    op_,
-                    self.wf.num_inactive_spin_orbs,
-                    self.wf.num_active_spin_orbs,
-                    self.wf.num_virtual_spin_orbs,
-                )
-                self.G_ops.append(ResponseOperator((i, j, k, l, m, n), (a, b, c, d, e, f), op))
-        for i, a in self.wf.kappa_idx:
-            op_ = 2 ** (-1 / 2) * epq_pauli(a, i, self.wf.num_spin_orbs, self.wf.num_elec)
-            op = convert_pauli_to_hybrid_form(
-                op_,
-                self.wf.num_inactive_spin_orbs,
-                self.wf.num_active_spin_orbs,
-                self.wf.num_virtual_spin_orbs,
-            )
-            self.q_ops.append(ResponseOperator((i), (a), op))
-
-        num_parameters = len(self.G_ops) + len(self.q_ops)
-        self.A = np.zeros((num_parameters, num_parameters))
-        self.B = np.zeros((num_parameters, num_parameters))
-        self.Sigma = np.zeros((num_parameters, num_parameters))
-        self.Delta = np.zeros((num_parameters, num_parameters))
-        H_1i_1a = convert_pauli_to_hybrid_form(
-            hamiltonian_pauli_1i_1a(
-                self.wf.h_ao,
-                self.wf.g_ao,
-                self.wf.c_trans,
-                self.wf.num_inactive_spin_orbs,
-                self.wf.num_active_spin_orbs,
-                self.wf.num_virtual_spin_orbs,
-                num_elec,
-            ),
-            self.wf.num_inactive_spin_orbs,
-            self.wf.num_active_spin_orbs,
-            self.wf.num_virtual_spin_orbs,
-        )
-        H_en = convert_pauli_to_hybrid_form(
-            energy_hamiltonian_pauli(
-                self.wf.h_ao,
-                self.wf.g_ao,
-                self.wf.c_trans,
-                self.wf.num_inactive_spin_orbs,
-                self.wf.num_active_spin_orbs,
-                self.wf.num_virtual_spin_orbs,
-                num_elec,
-            ),
-            self.wf.num_inactive_spin_orbs,
-            self.wf.num_active_spin_orbs,
-            self.wf.num_virtual_spin_orbs,
-        )
         rdms = ReducedDenstiyMatrix(
             self.wf.num_inactive_orbs,
             self.wf.num_active_orbs,
@@ -220,13 +70,13 @@ class LinearResponseUCC(LinearResponseBaseClass):
         grad = np.zeros(2 * len(self.G_ops))
         for i, op in enumerate(self.G_ops):
             grad[i] = expectation_value_hybrid_flow(
-                self.wf.state_vector, [H_1i_1a, op.operator], self.wf.state_vector
+                self.wf.state_vector, [self.H_1i_1a, op.operator], self.wf.state_vector
             ) - (
                 self.wf.energy_elec
                 * expectation_value_hybrid_flow(self.wf.state_vector, [op.operator], self.wf.state_vector)
             )
             grad[i + len(self.G_ops)] = expectation_value_hybrid_flow(
-                self.wf.state_vector, [op.operator.dagger, H_1i_1a], self.wf.state_vector
+                self.wf.state_vector, [op.operator.dagger, self.H_1i_1a], self.wf.state_vector
             ) - (
                 self.wf.energy_elec
                 * expectation_value_hybrid_flow(
@@ -265,11 +115,11 @@ class LinearResponseUCC(LinearResponseBaseClass):
                 GI = opI.operator
                 # Make A
                 self.A[j, i + idx_shift] = self.A[i + idx_shift, j] = expectation_value_hybrid_flow(
-                    self.wf.state_vector, [GI.dagger, H_1i_1a, qJ], self.wf.state_vector
+                    self.wf.state_vector, [GI.dagger, self.H_1i_1a, qJ], self.wf.state_vector
                 )
                 # Make B
                 self.B[j, i + idx_shift] = self.B[i + idx_shift, j] = -expectation_value_hybrid_flow(
-                    self.wf.state_vector, [GI.dagger, qJ.dagger, H_1i_1a], self.wf.state_vector
+                    self.wf.state_vector, [GI.dagger, qJ.dagger, self.H_1i_1a], self.wf.state_vector
                 )
         for j, opJ in enumerate(self.G_ops):
             GJ = opJ.operator
@@ -279,7 +129,7 @@ class LinearResponseUCC(LinearResponseBaseClass):
                     continue
                 # Make A
                 val = expectation_value_hybrid_flow(
-                    self.wf.state_vector, [GI.dagger, H_en, GJ], self.wf.state_vector
+                    self.wf.state_vector, [GI.dagger, self.H_en, GJ], self.wf.state_vector
                 )
                 val -= (
                     expectation_value_hybrid_flow(self.wf.state_vector, [GI.dagger, GJ], self.wf.state_vector)
@@ -287,7 +137,7 @@ class LinearResponseUCC(LinearResponseBaseClass):
                 )
                 val -= expectation_value_hybrid_flow(
                     self.wf.state_vector, [GI.dagger], self.wf.state_vector
-                ) * expectation_value_hybrid_flow(self.wf.state_vector, [H_en, GJ], self.wf.state_vector)
+                ) * expectation_value_hybrid_flow(self.wf.state_vector, [self.H_en, GJ], self.wf.state_vector)
                 val += (
                     expectation_value_hybrid_flow(self.wf.state_vector, [GI.dagger], self.wf.state_vector)
                     * expectation_value_hybrid_flow(self.wf.state_vector, [GJ], self.wf.state_vector)
@@ -296,7 +146,7 @@ class LinearResponseUCC(LinearResponseBaseClass):
                 self.A[i + idx_shift, j + idx_shift] = self.A[j + idx_shift, i + idx_shift] = val
                 # Make B
                 val = expectation_value_hybrid_flow(
-                    self.wf.state_vector, [GI.dagger, H_en], self.wf.state_vector
+                    self.wf.state_vector, [GI.dagger, self.H_en], self.wf.state_vector
                 ) * expectation_value_hybrid_flow(self.wf.state_vector, [GJ.dagger], self.wf.state_vector)
                 val -= (
                     expectation_value_hybrid_flow(self.wf.state_vector, [GI.dagger], self.wf.state_vector)
